+++ integration-tests/api/helpers/cart-seeder.js
@@ -14,31 +14,47 @@ const {
   LineItem,
   Payment,
   PaymentSession,
-} = require("@medusajs/medusa");
+} = require("@medusajs/medusa")
 
 module.exports = async (connection, data = {}) => {
-  const manager = connection.manager;
+  const manager = connection.manager
 
   const defaultProfile = await manager.findOne(ShippingProfile, {
     type: "default",
-  });
+  })
 
   const gcProfile = await manager.findOne(ShippingProfile, {
     type: "gift_card",
-  });
+  })
 
   await manager.insert(Address, {
     id: "test-general-address",
     first_name: "superman",
     country_code: "us",
-  });
+  })
 
   const r = manager.create(Region, {
     id: "test-region",
     name: "Test Region",
     currency_code: "usd",
     tax_rate: 0,
-  });
+  })
+
+  // Region with multiple countries
+  const regionWithMultipleCoutries = manager.create(Region, {
+    id: "test-region-multiple",
+    name: "Test Region",
+    currency_code: "eur",
+    tax_rate: 0,
+  })
+
+  await manager.save(regionWithMultipleCoutries)
+  await manager.query(
+    `UPDATE "country" SET region_id='test-region-multiple' WHERE iso_2 = 'no'`
+  )
+  await manager.query(
+    `UPDATE "country" SET region_id='test-region-multiple' WHERE iso_2 = 'dk'`
+  )
 
   const freeRule = manager.create(DiscountRule, {
     id: "free-shipping-rule",
@@ -46,18 +62,18 @@ module.exports = async (connection, data = {}) => {
     type: "free_shipping",
     value: 100,
     allocation: "total",
-  });
+  })
 
   const freeDisc = manager.create(Discount, {
     id: "free-shipping",
     code: "FREE_SHIPPING",
     is_dynamic: false,
     is_disabled: false,
-  });
+  })
 
-  freeDisc.regions = [r];
-  freeDisc.rule = freeRule;
-  await manager.save(freeDisc);
+  freeDisc.regions = [r]
+  freeDisc.rule = freeRule
+  await manager.save(freeDisc)
 
   const tenPercentRule = manager.create(DiscountRule, {
     id: "tenpercent-rule",
@@ -65,25 +81,25 @@ module.exports = async (connection, data = {}) => {
     type: "percentage",
     value: 10,
     allocation: "total",
-  });
+  })
 
   const tenPercent = manager.create(Discount, {
     id: "10Percent",
     code: "10PERCENT",
     is_dynamic: false,
     is_disabled: false,
-  });
+  })
 
-  tenPercent.regions = [r];
-  tenPercent.rule = tenPercentRule;
-  await manager.save(tenPercent);
+  tenPercent.regions = [r]
+  tenPercent.rule = tenPercentRule
+  await manager.save(tenPercent)
 
   const d = await manager.create(Discount, {
     id: "test-discount",
     code: "CREATED",
     is_dynamic: false,
     is_disabled: false,
-  });
+  })
 
   const dr = await manager.create(DiscountRule, {
     id: "test-discount-rule",
@@ -91,31 +107,31 @@ module.exports = async (connection, data = {}) => {
     type: "fixed",
     value: 10000,
     allocation: "total",
-  });
+  })
 
-  d.rule = dr;
-  d.regions = [r];
+  d.rule = dr
+  d.regions = [r]
 
-  await manager.save(d);
+  await manager.save(d)
 
   await manager.query(
     `UPDATE "country" SET region_id='test-region' WHERE iso_2 = 'us'`
-  );
+  )
 
   await manager.insert(Customer, {
     id: "test-customer",
     email: "test@email.com",
-  });
+  })
 
   await manager.insert(Customer, {
     id: "test-customer-2",
     email: "test-2@email.com",
-  });
+  })
 
   await manager.insert(Customer, {
     id: "some-customer",
     email: "some-customer@email.com",
-  });
+  })
 
   await manager.insert(ShippingOption, {
     id: "test-option",
@@ -126,7 +142,7 @@ module.exports = async (connection, data = {}) => {
     price_type: "flat_rate",
     amount: 1000,
     data: {},
-  });
+  })
 
   await manager.insert(ShippingOption, {
     id: "gc-option",
@@ -137,7 +153,7 @@ module.exports = async (connection, data = {}) => {
     price_type: "flat_rate",
     amount: 0,
     data: {},
-  });
+  })
 
   await manager.insert(ShippingOption, {
     id: "test-option-2",
@@ -148,7 +164,7 @@ module.exports = async (connection, data = {}) => {
     price_type: "flat_rate",
     amount: 500,
     data: {},
-  });
+  })
 
   await manager.insert(Product, {
     id: "giftcard-product",
@@ -157,7 +173,7 @@ module.exports = async (connection, data = {}) => {
     discountable: false,
     profile_id: gcProfile.id,
     options: [{ id: "denom", title: "Denomination" }],
-  });
+  })
 
   await manager.insert(ProductVariant, {
     id: "giftcard-denom",
@@ -170,14 +186,14 @@ module.exports = async (connection, data = {}) => {
         value: "1000",
       },
     ],
-  });
+  })
 
   await manager.insert(Product, {
     id: "test-product",
     title: "test product",
     profile_id: defaultProfile.id,
     options: [{ id: "test-option", title: "Size" }],
-  });
+  })
 
   await manager.insert(ProductVariant, {
     id: "test-variant",
@@ -190,7 +206,7 @@ module.exports = async (connection, data = {}) => {
         value: "Size",
       },
     ],
-  });
+  })
 
   await manager.insert(ProductVariant, {
     id: "test-variant-2",
@@ -203,31 +219,31 @@ module.exports = async (connection, data = {}) => {
         value: "Size",
       },
     ],
-  });
+  })
 
   const ma = manager.create(MoneyAmount, {
     variant_id: "test-variant",
     currency_code: "usd",
     amount: 1000,
-  });
+  })
 
-  await manager.save(ma);
+  await manager.save(ma)
 
   const ma2 = manager.create(MoneyAmount, {
     variant_id: "test-variant-2",
     currency_code: "usd",
     amount: 8000,
-  });
+  })
 
-  await manager.save(ma2);
+  await manager.save(ma2)
 
   const ma3 = manager.create(MoneyAmount, {
     variant_id: "giftcard-denom",
     currency_code: "usd",
     amount: 1000,
-  });
+  })
 
-  await manager.save(ma3);
+  await manager.save(ma3)
 
   const cart = manager.create(Cart, {
     id: "test-cart",
@@ -241,9 +257,9 @@ module.exports = async (connection, data = {}) => {
     region_id: "test-region",
     currency_code: "usd",
     items: [],
-  });
+  })
 
-  await manager.save(cart);
+  await manager.save(cart)
 
   const cart2 = manager.create(Cart, {
     id: "test-cart-2",
@@ -258,7 +274,26 @@ module.exports = async (connection, data = {}) => {
     currency_code: "usd",
     completed_at: null,
     items: [],
-  });
+  })
+
+  const swapCart = manager.create(Cart, {
+    id: "swap-cart",
+    type: "swap",
+    customer_id: "some-customer",
+    email: "some-customer@email.com",
+    shipping_address: {
+      id: "test-shipping-address",
+      first_name: "lebron",
+      country_code: "us",
+    },
+    region_id: "test-region",
+    currency_code: "usd",
+    completed_at: null,
+    items: [],
+    metadata: {
+      swap_id: "test-swap",
+    },
+  })
 
   const pay = manager.create(Payment, {
     id: "test-payment",
@@ -267,13 +302,25 @@ module.exports = async (connection, data = {}) => {
     amount_refunded: 0,
     provider_id: "test-pay",
     data: {},
-  });
+  })
 
-  await manager.save(pay);
+  const swapPay = manager.create(Payment, {
+    id: "test-swap-payment",
+    amount: 10000,
+    currency_code: "usd",
+    amount_refunded: 0,
+    provider_id: "test-pay",
+    data: {},
+  })
 
-  cart2.payment = pay;
+  await manager.save(pay)
+  await manager.save(swapPay)
 
-  await manager.save(cart2);
+  cart2.payment = pay
+  swapCart.payment = swapPay
+
+  await manager.save(cart2)
+  await manager.save(swapCart)
 
   await manager.insert(PaymentSession, {
     id: "test-session",
@@ -282,7 +329,16 @@ module.exports = async (connection, data = {}) => {
     is_selected: true,
     data: {},
     status: "authorized",
-  });
+  })
+
+  await manager.insert(PaymentSession, {
+    id: "test-swap-session",
+    cart_id: "swap-cart",
+    provider_id: "test-pay",
+    is_selected: true,
+    data: {},
+    status: "authorized",
+  })
 
   await manager.insert(ShippingMethod, {
     id: "test-method",
@@ -290,7 +346,7 @@ module.exports = async (connection, data = {}) => {
     cart_id: "test-cart",
     price: 1000,
     data: {},
-  });
+  })
 
   const li = manager.create(LineItem, {
     id: "test-item",
@@ -301,8 +357,8 @@ module.exports = async (connection, data = {}) => {
     quantity: 1,
     variant_id: "test-variant",
     cart_id: "test-cart-2",
-  });
-  await manager.save(li);
+  })
+  await manager.save(li)
 
   const cart3 = manager.create(Cart, {
     id: "test-cart-3",
@@ -317,8 +373,8 @@ module.exports = async (connection, data = {}) => {
     currency_code: "usd",
     completed_at: null,
     items: [],
-  });
-  await manager.save(cart3);
+  })
+  await manager.save(cart3)
 
   await manager.insert(ShippingMethod, {
     id: "test-method-2",
@@ -326,7 +382,7 @@ module.exports = async (connection, data = {}) => {
     cart_id: "test-cart-3",
     price: 0,
     data: {},
-  });
+  })
 
   const li2 = manager.create(LineItem, {
     id: "test-item-2",
@@ -337,6 +393,6 @@ module.exports = async (connection, data = {}) => {
     quantity: 1,
     variant_id: "test-variant",
     cart_id: "test-cart-3",
-  });
-  await manager.save(li2);
-};
+  })
+  await manager.save(li2)
+}