+++ packages/medusa/src/api/routes/store/products/__tests__/list-products.js
@@ -1,4 +1,5 @@
 import { IdMap } from "medusa-test-utils"
+import { defaultRelations } from ".."
 import { request } from "../../../../../helpers/test-request"
 import { ProductServiceMock } from "../../../../../services/__mocks__/product"
 
@@ -17,8 +18,8 @@ describe("GET /store/products", () => {
     it("calls get product from productSerice", () => {
       expect(ProductServiceMock.list).toHaveBeenCalledTimes(1)
       expect(ProductServiceMock.list).toHaveBeenCalledWith(
-        {},
-        { relations: ["variants", "options", "images"], skip: 0, take: 100 }
+        { status: ["published"] },
+        { relations: defaultRelations, skip: 0, take: 100 }
       )
     })
 
@@ -42,8 +43,8 @@ describe("GET /store/products", () => {
     it("calls list from productSerice", () => {
       expect(ProductServiceMock.list).toHaveBeenCalledTimes(1)
       expect(ProductServiceMock.list).toHaveBeenCalledWith(
-        { is_giftcard: true },
-        { relations: ["variants", "options", "images"], skip: 0, take: 100 }
+        { is_giftcard: true, status: ["published"] },
+        { relations: defaultRelations, skip: 0, take: 100 }
       )
     })
   })