+++ packages/medusa/src/services/__mocks__/swap.js
@@ -11,6 +11,21 @@ export const SwapServiceMock = {
     return Promise.resolve()
   }),
   retrieve: jest.fn().mockImplementation(data => {
+    switch (data) {
+      case IdMap.getId("test-swap"):
+        return Promise.resolve({
+          id: "test-swap",
+          order_id: IdMap.getId("test-order"),
+        })
+      default:
+        return Promise.resolve({ id: "test-swap" })
+    }
+  }),
+  cancel: jest.fn().mockImplementation(f => {
+    return Promise.resolve({ f })
+  }),
+
+  cancelFulfillment: jest.fn().mockImplementation(f => {
     return Promise.resolve({ id: IdMap.getId("test-swap") })
   }),
   list: jest.fn().mockImplementation((...args) => {