+++ packages/medusa/src/services/__tests__/product.js
@@ -11,9 +11,26 @@ const eventBusService = {
 describe("ProductService", () => {
   describe("retrieve", () => {
     const productRepo = MockRepository({
-      findOneWithRelations: () =>
-        Promise.resolve({ id: IdMap.getId("ironman") }),
+      findOneWithRelations: (rels, query) => {
+        if (query.where.id === "test id with variants") {
+          return {
+            id: "test id with variants",
+            variants: [
+              { id: "test_321", title: "Green" },
+              { id: "test_123", title: "Blue" },
+            ],
+          }
+        }
+        if (query.where.id === "test id one variant") {
+          return {
+            id: "test id one variant",
+            variants: [{ id: "test_123", title: "Blue" }],
+          }
+        }
+        return Promise.resolve({ id: IdMap.getId("ironman") })
+      },
     })
+
     const productService = new ProductService({
       manager: MockManager,
       productRepository: productRepo,
@@ -37,11 +54,12 @@ describe("ProductService", () => {
 
   describe("create", () => {
     const productRepository = MockRepository({
-      create: () => ({
+      create: product => ({
         id: IdMap.getId("ironman"),
         title: "Suit",
         options: [],
         collection: { id: IdMap.getId("cat"), title: "Suits" },
+        variants: product.variants,
       }),
       findOneWithRelations: () => ({
         id: IdMap.getId("ironman"),
@@ -97,6 +115,16 @@ describe("ProductService", () => {
         options: [],
         tags: [{ value: "title" }, { value: "title2" }],
         type: "type-1",
+        variants: [
+          {
+            id: "test1",
+            title: "green",
+          },
+          {
+            id: "test2",
+            title: "blue",
+          },
+        ],
       })
 
       expect(eventBusService.emit).toHaveBeenCalledTimes(1)
@@ -108,6 +136,16 @@ describe("ProductService", () => {
       expect(productRepository.create).toHaveBeenCalledTimes(1)
       expect(productRepository.create).toHaveBeenCalledWith({
         title: "Suit",
+        variants: [
+          {
+            id: "test1",
+            title: "green",
+          },
+          {
+            id: "test2",
+            title: "blue",
+          },
+        ],
       })
 
       expect(productTagRepository.findOne).toHaveBeenCalledTimes(2)
@@ -124,14 +162,30 @@ describe("ProductService", () => {
         title: "Suit",
         options: [],
         tags: [
-          { id: "tag-1", value: "title" },
-          { id: "tag-2", value: "title2" },
+          {
+            id: "tag-1",
+            value: "title",
+          },
+          {
+            id: "tag-2",
+            value: "title2",
+          },
         ],
         type_id: "type",
         collection: {
           id: IdMap.getId("cat"),
           title: "Suits",
         },
+        variants: [
+          {
+            id: "test1",
+            title: "green",
+          },
+          {
+            id: "test2",
+            title: "blue",
+          },
+        ],
       })
     })
   })
@@ -145,8 +199,23 @@ describe("ProductService", () => {
             variants: [{ id: IdMap.getId("green"), title: "Green" }],
           })
         }
+        if (query.where.id === "prod_status") {
+          return Promise.resolve({
+            id: "prod_status",
+            status: "draft",
+          })
+        }
         if (query.where.id === "123") {
           return undefined
+        }
+        if (query.where.id === "ranking test") {
+          return Promise.resolve({
+            id: "ranking test",
+            variants: [
+              { id: "test_321", title: "Greener", variant_rank: 1 },
+              { id: "test_123", title: "Blueer", variant_rank: 0 },
+            ],
+          })
         }
         return Promise.resolve({ id: IdMap.getId("ironman") })
       },
@@ -165,7 +234,12 @@ describe("ProductService", () => {
       withTransaction: function() {
         return this
       },
-      update: () => Promise.resolve(),
+      update: (variant, update) => {
+        if (variant.id) {
+          return update
+        }
+        return Promise.resolve()
+      },
     }
 
     const productTagRepository = MockRepository({
@@ -222,6 +296,18 @@ describe("ProductService", () => {
       expect(productRepository.save).toHaveBeenCalledTimes(1)
     })
 
+    it("successfully updates product status", async () => {
+      await productService.update(IdMap.getId("ironman"), {
+        status: "published",
+      })
+
+      expect(productRepository.save).toHaveBeenCalledTimes(1)
+      expect(productRepository.save).toHaveBeenCalledWith({
+        id: IdMap.getId("ironman"),
+        status: "published",
+      })
+    })
+
     it("successfully updates product", async () => {
       await productService.update(IdMap.getId("ironman"), {
         title: "Full suit",
@@ -245,6 +331,30 @@ describe("ProductService", () => {
           id: IdMap.getId("test"),
           value: "test",
         },
+      })
+    })
+
+    it("successfully updates variant ranking", async () => {
+      await productService.update("ranking test", {
+        variants: [
+          { id: "test_321", title: "Greener", variant_rank: 1 },
+          { id: "test_123", title: "Blueer", variant_rank: 0 },
+        ],
+      })
+
+      expect(eventBusService.emit).toHaveBeenCalledTimes(1)
+      expect(eventBusService.emit).toHaveBeenCalledWith(
+        "product.updated",
+        expect.any(Object)
+      )
+
+      expect(productRepository.save).toHaveBeenCalledTimes(1)
+      expect(productRepository.save).toHaveBeenCalledWith({
+        id: "ranking test",
+        variants: [
+          { id: "test_321", title: "Greener", variant_rank: 0 },
+          { id: "test_123", title: "Blueer", variant_rank: 1 },
+        ],
       })
     })