+++ packages/medusa/src/services/payment-provider.js
@@ -321,7 +321,6 @@ class PaymentProviderService extends BaseService {
   async cancelPayment(paymentObj) {
     return this.atomicPhase_(async manager => {
       const payment = await this.retrievePayment(paymentObj.id)
-
       const provider = this.retrieveProvider(payment.provider_id)
       payment.data = await provider.cancelPayment(payment)
 
@@ -329,7 +328,7 @@ class PaymentProviderService extends BaseService {
       payment.canceled_at = now.toISOString()
 
       const paymentRepo = manager.getCustomRepository(this.paymentRepository_)
-      return paymentRepo.save(payment)
+      return await paymentRepo.save(payment)
     })
   }